Moved createEvent modal to new file
